+++ packages/web-components/components/smart-camera-web/src/SmartCameraWeb.js
@@ -67,7 +67,6 @@ class SmartCameraWeb extends HTMLElement {
       'hide-back-of-id',
       'hide-back-to-host',
       'show-navigation',
-      'hide-back-of-id',
       'theme-color',
       'hide-attribution',
     ];
@@ -82,6 +81,7 @@ class SmartCameraWeb extends HTMLElement {
       case 'hide-back-of-id':
       case 'hide-back-to-host':
       case 'show-navigation':
+      case 'theme-color':
         this.disconnectedCallback();
         this.shadowRoot.innerHTML = this.render();
         this.setUpEventListeners();