+++ packages/web-components/components/totp-consent/src/TotpConsent.js
@@ -196,7 +196,7 @@ function markup() {
             .back-button-text {
                 font-size: 11px;
                 line-height: 11px;
-                color: rgb(21, 31, 114);
+                color: ${this.themeColor || 'rgb(21, 31, 114)'};
             }
 
             #error,
@@ -290,7 +290,7 @@ function markup() {
                     <button type='button' data-type='icon' id="back-button" class="back-button">
                         <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none">
                             <path fill="#DBDBC4" d="M12 22c5.523 0 10-4.477 10-10S17.523 2 12 2 2 6.477 2 12s4.477 10 10 10Z" opacity=".4"/>
-                            <path fill="#001096" d="M15.5 11.25h-5.19l1.72-1.72c.29-.29.29-.77 0-1.06a.754.754 0 0 0-1.06 0l-3 3c-.29.29-.29.77 0 1.06l3 3c.15.15.34.22.53.22s.38-.07.53-.22c.29-.29.29-.77 0-1.06l-1.72-1.72h5.19c.41 0 .75-.34.75-.75s-.34-.75-.75-.75Z"/>
+                            <path fill="${this.themeColor}" d="M15.5 11.25h-5.19l1.72-1.72c.29-.29.29-.77 0-1.06a.754.754 0 0 0-1.06 0l-3 3c-.29.29-.29.77 0 1.06l3 3c.15.15.34.22.53.22s.38-.07.53-.22c.29-.29.29-.77 0-1.06l-1.72-1.72h5.19c.41 0 .75-.34.75-.75s-.34-.75-.75-.75Z"/>
                         </svg>
                     </button>
                     <div class="back-button-text">Back</div>
@@ -337,7 +337,7 @@ function markup() {
                     <button type='button' data-type='icon' id="back-to-entry-button" class="back-button">
                         <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none">
                             <path fill="#DBDBC4" d="M12 22c5.523 0 10-4.477 10-10S17.523 2 12 2 2 6.477 2 12s4.477 10 10 10Z" opacity=".4"/>
-                            <path fill="#001096" d="M15.5 11.25h-5.19l1.72-1.72c.29-.29.29-.77 0-1.06a.754.754 0 0 0-1.06 0l-3 3c-.29.29-.29.77 0 1.06l3 3c.15.15.34.22.53.22s.38-.07.53-.22c.29-.29.29-.77 0-1.06l-1.72-1.72h5.19c.41 0 .75-.34.75-.75s-.34-.75-.75-.75Z"/>
+                            <path fill="${this.themeColor}" d="M15.5 11.25h-5.19l1.72-1.72c.29-.29.29-.77 0-1.06a.754.754 0 0 0-1.06 0l-3 3c-.29.29-.29.77 0 1.06l3 3c.15.15.34.22.53.22s.38-.07.53-.22c.29-.29.29-.77 0-1.06l-1.72-1.72h5.19c.41 0 .75-.34.75-.75s-.34-.75-.75-.75Z"/>
                         </svg>
                     </button>
                     <div class="back-button-text">Back</div>
@@ -891,6 +891,18 @@ class TotpConsent extends HTMLElement {
 
   get token() {
     return this.getAttribute('token');
+  }
+
+  get themeColor() {
+    return this.getAttribute('theme-color') || '#001096';
+  }
+
+  get hideBack() {
+    return this.hasAttribute('hide-back');
+  }
+
+  get showNavigation() {
+    return this.hasAttribute('show-navigation');
   }
 
   handleTotpConsentGrant() {